+++ NEWS.md
@@ -1,4 +1,18 @@
+# WP6RTools 0.0.2
+
+- Added `oecd_secretary()` a function scraping the table of Secretary Generals
+of the OECD from Wikipedia
+- Added `oecd_members()` a function scraping the table of member countries
+of the OECD from Wikipedia
+- Added `oecd_member_metrics()` a function scraping the table of member
+countries of the OECD and their socioeconomic characteristics from Wikipedia
+
 # WP6RTools 0.0.1
+
+- Added `theme_oecd()` and `theme_oecd_dark()`, two themes for `{ggplot}` graphs
+- Added `oecd_palettes()` a list of oecd palettes
+- Added `oecd_palette()` a function to generate discrete and continuous palettes
+based on the OECD colors
 
 # WP6RTools 0.0.0.9000